META-4203 Fix Keycloak issues (#1782)
* META-4203 Add keycloak Admin client

* META-4203 Add keycloak Admin client

* META-4203: TypeDefs AuthPolicy, AuthService

* META-4203: Use Atlas authorizer with configuration in peoperties file

* META-4203: PolicyRefresher to fetch roles,users cache from keycloak

* META-4203: PolicyRefresher to fetch roles,users cache from keycloak

* META-4203: PolicyRefresher APIs to fetch users, roles

* META-4203: Change policyType from int to String

* META-4203: Change policyType from int to String

* META-4203 Policy transformer for caching

* META-4203 Policy transformer for caching

* META-4203 Policy transformer for caching

* META-4203 Policy transformer for caching

* META-4203 Policy transformer for caching

* META-4203 Policy transformer: serviceId to String from int

* META-4203 ES access loggin into index ranger-audit

* META-4203 ES access loggin into index ranger-audit

* META-4203 Fix review comment

* META-4203 Address review comment

* META-4203 Address review comment

* META-4203 Authpolicy & AuthService bootstrapper

* META-4203 PreProcessor for Persona, Purpose, AuthPolicy

* META-4203 PreProcessor for Connection, Collection

* META-4203 Fix github code scanning problem

* Create MigrationREST.java

* Update Security config Paths

* No need to open migration rest APIs

* fix: Update API endpoints

* META-4203 Create types conditionally

* META-4203 SkipScrubbing true for Persona users, groups

* META-4203 Fix: Persona/Purpose ES alias filters

* META-4203 Fix: Update keycloak role issue while updating Conn/Coll/Persona

* META-4203 Address review comments

* META-4203 Address Review comments

* META-4203 Keycloak client fix issues

* META-4203 Sync with changes made in killRanger

* META-4203 Fix Keycloak issues

* META-4203 Allow migration without Atlas authz property

---------

(cherry picked from commit de1441226dadffec4829e354f533100c46c61071)

+++ repository/src/main/java/org/apache/atlas/repository/store/graph/v2/preprocessor/sql/QueryCollectionPreProcessor.java
@@ -161,20 +161,22 @@ public class QueryCollectionPreProcessor implements PreProcessor {
             AtlasEntity.AtlasEntityWithExtInfo entityWithExtInfo = entityRetriever.toAtlasEntityWithExtInfo(vertex);
             AtlasEntity collection = entityWithExtInfo.getEntity();
 
-            String adminRoleName = String.format(COLL_ADMIN_ROLE_PATTERN, collection.getGuid());
-            String viewerRoleName = String.format(COLL_VIEWER_ROLE_PATTERN, collection.getGuid());
-
             if (!AtlasEntity.Status.ACTIVE.equals(collection.getStatus())) {
                 throw new AtlasBaseException("Collection is already deleted/purged");
             }
 
-            //delete connection policies
-            List<AtlasEntityHeader> policies = getCollectionPolicies(collection.getGuid());
-            EntityMutationResponse response = entityStore.deleteByIds(policies.stream().map(x -> x.getGuid()).collect(Collectors.toList()));
+            if (ATLAS_AUTHORIZER_IMPL.equalsIgnoreCase(CURRENT_AUTHORIZER_IMPL)) {
+                //delete collection policies
+                List<AtlasEntityHeader> policies = getCollectionPolicies(collection.getGuid());
+                EntityMutationResponse response = entityStore.deleteByIds(policies.stream().map(x -> x.getGuid()).collect(Collectors.toList()));
 
-            //delete connection role
-            keycloakStore.removeRoleByName(adminRoleName);
-            keycloakStore.removeRoleByName(viewerRoleName);
+                //delete collection role
+                String adminRoleName = String.format(COLL_ADMIN_ROLE_PATTERN, collection.getGuid());
+                String viewerRoleName = String.format(COLL_VIEWER_ROLE_PATTERN, collection.getGuid());
+
+                keycloakStore.removeRoleByName(adminRoleName);
+                keycloakStore.removeRoleByName(viewerRoleName);
+            }
         } finally {
             RequestContext.get().endMetricRecord(metricRecorder);
         }
